exclude walk time from transit directions
walk times added a huge amount of variability to public transit
durations. You can interpret this choice to ignore walk time as "we'll
choose to live near a bus stop or train station", and most routes have
only a few minutes of walking between transit legs and at the end.